requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
ci-info==0.2.0
click==8.0.3
configobj==5.0.6
configparser==5.2.0
cryptography==36.0.1
Django==4.0.2
djangorestframework==3.13.1
etelemetry==0.3.0
filelock==3.4.2
fitz==0.0.1.dev2
future==0.18.2
google-api-core==2.5.0
google-auth==2.6.0
google-cloud==0.34.0
google-cloud-core==2.2.2
google-cloud-speech==2.12.0
google-cloud-storage==2.1.0
google-cloud-vision==2.6.3
google-crc32c==1.3.0
google-resumable-media==2.2.1
googleapis-common-protos==1.54.0
grpcio==1.43.0
grpcio-status==1.43.0
httplib2==0.20.4
idna==3.3
isodate==0.6.1
libcst==0.4.1
lxml==4.7.1
mypy-extensions==0.4.3
networkx==2.6.3
nibabel==3.2.2
nipype==1.7.0
numpy==1.22.2
packaging==21.3
pandas==1.4.1
pathlib==1.0.1
pdf2image==1.16.0
pdfminer.six==20211012
Pillow==9.0.1
proto-plus==1.20.0
protobuf==3.19.4
prov==2.0.0
<<<<<<< HEAD
pycparser==2.21
=======
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.14.1
PyDocX==0.9.10
>>>>>>> 0b9fe8a2ad3f27906c77637df32f0d38f0170837
pydot==1.4.2
pydub==0.25.1
PyMuPDF==1.19.5
pyocr==0.8.1
pyparsing==3.0.7
PyPDF2==1.26.0
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
pytesseract==0.3.8
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.19.2
pytz==2021.3
pyxnat==1.4
PyYAML==6.0
rdflib==6.1.1
requests==2.27.1
rsa==4.8
scipy==1.8.0
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
tika==1.24
traits==6.3.2
typing-extensions==4.1.1
typing-inspect==0.7.1
urllib3==1.26.8
